edit: update_method/model / add: jsonloader
