+++ requirements.txt
@@ -24,3 +24,4 @@ tqdm
 # used for application
 pygame
 PySide6
+playsound
