+++ src/application/WavePlayer.py
@@ -1,0 +1,1 @@
+# pygameのmixerモジュールを使ってwavファイルを再生
